Added a power up
When the power up is picked up by the player, it will fire the assigned event and then destroy itself.
Added a script that calls an event at a set interval. I use this to spawn a power up where the ball is every 3 seconds but it's pretty generic.
Added a simple script that provides a public function to spawn a gameobject that is assigned in the inspector. You can call this public function from the inspector to allow other components to spawn objects when an event is fired.